+++ src/htdocs/js/NavBar.js
@@ -166,6 +166,10 @@ var NavBar = function (options) {
     // Update map container / render plots so they display correctly when unhidden
     if (id === 'mapPane') {
       _MapPane.map.invalidateSize();
+      // Fire an event so L.popup.update() can be called after map is visible,
+      // which seems to be necessary for Leaflet to display popups correctly
+      // when they're added to the map from another pane
+      _MapPane.map.fire('visible');
     } else if (id === 'plotsPane') {
       _PlotsPane.renderPlots();
       _PlotsPane.resizePlots();